Option: Modifier objet servlet+jsp

+++ src/main/java/com/epf/rentmanager/model/Vehicle.java
@@ -16,7 +16,7 @@ public class Vehicle {
         this.nb_places = nb_places;
     }
 
-    public Vehicle(int id, String constructeur, String modele, int nb_places) {
+    public Vehicle(long id, String constructeur, String modele, int nb_places) {
         this.id = id;
         this.constructeur = constructeur;
         this.modele = modele;
